// sanity.types.ts
export type Link = {
  _type: "link";
  linkType?: "href" | "page" | "post" | "article" | "slug";
  href?: string;
  page?: {
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: "page";
  };
  post?: {
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: "post";
  };
  article?: {
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: "article";
  };
  slug?: string;
  openInNewTab?: boolean;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;